agents/requirements.txt
agno>=0.1.0
python-dotenv>=1.0.0
chromadb>=0.4.0
google-generativeai>=0.3.0
anthropic>=0.5.0
httpx>=0.24.0
pydantic>=2.0.0 